requirements.txt
aniso8601==9.0.1
click==8.0.3
Flask==2.0.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2-binary==2.9.2
PyJWT==1.7.1
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.27
Werkzeug==2.0.2
